+++ tests/test_windows.py
@@ -3,7 +3,6 @@ This is part of the MSS Python's module.
 Source: https://github.com/BoboTiG/python-mss
 """
 
-import ctypes.util
 import platform
 
 import mss
@@ -18,6 +17,6 @@ if platform.system().lower() != "windows":
 def test_implementation(monkeypatch):
     # Test bad data retrieval
     with mss.mss() as sct:
-        monkeypatch.setattr(ctypes.windll.gdi32, "GetDIBits", lambda *args: 0)
+        monkeypatch.setattr(sct.gdi32, "GetDIBits", lambda *args: 0)
         with pytest.raises(ScreenShotError):
             sct.shot()